refactor: Move city-related controllers to the Cities directory.
