+++ app/Http/Controllers/Cities/DestroyCityController.php
@@ -1,13 +1,13 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\Cities;
 
 use App\Models\City;
 use Flugg\Responder\Contracts\Responder;
 use Illuminate\Database\Eloquent\ModelNotFoundException;
 use Illuminate\Http\JsonResponse;
 
-class DestroyCityController extends Controller
+class DestroyCityController
 {
     public function __invoke(int $id, Responder $responder): JsonResponse
     {
