--- app/Http/Controllers/Controller.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace App\Http\Controllers;
-
-abstract class Controller
-{
-}
